% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{binned_mean_regprob}
\alias{binned_mean_regprob}
\title{Calculate mean regulation probabilities for binned genes for each component}
\usage{
binned_mean_regprob(regprobs = regprobs_matrix_custom, nbins = 50,
  factorisation = SDAresults)
}
\arguments{
\item{regprobs}{matrix; rows = genes, columns = motif/transcription factors}

\item{nbins}{integer; number of bins to put gene loadings into (equal number of genes in each)}

\item{factorisation}{list; output from SDAtools::load_results()}
}
\value{
3d Array / Tensor, of motif by bin by component
}
\description{
Calculate mean regulation probabilities for binned genes for each component
}
