% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{clustered_heatmap}
\alias{clustered_heatmap}
\title{Clustered heatmap (partially deprecated)}
\usage{
clustered_heatmap(cell_metadata = cell_data, name,
  annotation.col = NULL, colv_order = NULL, col_lab = NULL,
  row_text_size = 0.5)
}
\arguments{
\item{cell_metadata}{data.table; cell_data, with subset of gene columns}

\item{name}{string; title}

\item{annotation.col}{data.table; subset of cell_data, passed to annCol of aheatmap}

\item{colv_order}{passed to Colv of aheatmap}

\item{col_lab}{passed to labCol of aheatmap}

\item{row_text_size}{passed to cexRow of aheatmap}
}
\value{
aheatmap plot
}
\description{
Clustered heatmap (partially deprecated)
}
\details{
requires PCA object to be bound to cell_metadata
}
