% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{compare_factorisations}
\alias{compare_factorisations}
\title{Plot correlation heatmap comparing two sets of loadings}
\usage{
compare_factorisations(f1 = nnmf_decomp$H,
  f2 = SDAresults$loadings[[1]], names = c("f1", "f2"),
  method = "spearman", return = "hm", split = c(F, F),
  inject_matrix = NULL, randomise = FALSE)
}
\arguments{
\item{f1}{gene loadings from factorisation method 1}

\item{f2}{gene loadings from factorisation method 1}

\item{names}{string; names of f1 and f2 used in heatmap}

\item{method}{correlation method, note kendall not allowed}

\item{return_cor}{logical; should the ordered correlation matrix be returned,
if FALSE (default) heatmap is plotted using ComplexHeatmap package}
}
\value{
ggplot2 object
}
\description{
Plot correlation heatmap comparing two sets of loadings
}
