% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{correlate_regprobs}
\alias{correlate_regprobs}
\title{For a matrix of genes by motifs, calculate correlations}
\usage{
correlate_regprobs(regprobs = regprobs_matrix,
  factorisation = SDAresults)
}
\arguments{
\item{regprobs}{matrix; a matrix of genes by motifs}

\item{factorisation}{SDA factorisation object, output of SDAtools::load_results()}
}
\value{
Matrix of components by motifs, with entries for the test statistic t for each combination
}
\description{
For a matrix of genes by motifs, calculate correlations
}
\details{
Calculates two-sided Pearson's product-moment correlation test. Positive and negative loadings are split and calculated seperately.
}
