% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{create_grob_list}
\alias{create_grob_list}
\title{Create list of grobs from a function and input}
\usage{
create_grob_list(fn = print_marker2, input = hist_subset)
}
\arguments{
\item{fn;}{function which returns ggplot object}

\item{input;}{vector to be fed to the function}
}
\description{
Create list of grobs from a function and input
}
\details{
Maybe this is redundant with lapply?
}
