% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{dmultinom_v}
\alias{dmultinom_v}
\title{Vectorised multionimial distribution}
\usage{
dmultinom_v(counts, prob)
}
\arguments{
\item{counts}{numeric matrix; matrix of K columns of integers in 0:size. each sample is a row}

\item{prob}{numeric non-negative matrix with K columns, specifying the probability for the K classes; is internally normalized to sum 1. Infinite and missing values are not allowed. Each sample is a row.}
}
\value{
LOG probability density / likelihood of the data counts given the parameters in prob
}
\description{
Vectorised multionimial distribution
}
\details{
vectorised version of R base dmultinom
}
