% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{download_flank_sequences}
\alias{download_flank_sequences}
\title{Download sequences flanking gene}
\usage{
download_flank_sequences(gene_names, length, flank_type = "gene_flank")
}
\description{
Download sequences flanking gene
}
\details{
This function is deprecated, use bedtools getfasta instead
}
