% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{download_tss}
\alias{download_tss}
\title{Download location of Transcripion Start Sites from Ensembl Biomart}
\usage{
download_tss(gene_names)
}
\arguments{
\item{gene_names}{charachter vector; vector of gene names (symbols) to download TSS for}
}
\value{

}
\description{
Download location of Transcripion Start Sites from Ensembl Biomart
}
\details{
Uses ensembl biomart to download location of Transcripion Start Sites
If there are multiple tss the best is determined by tsl, then by appris, then by length.
}
