% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{export_FASTA}
\alias{export_FASTA}
\title{Export sequences to file in FASTA format}
\usage{
export_FASTA(sequences, file)
}
\arguments{
\item{sequences}{data.frame; data.table / data.frame}
}
\value{

}
\description{
Export sequences to file in FASTA format
}
\details{
Deprecated
}
