% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{find_motifs_parallel}
\alias{find_motifs_parallel}
\title{For known motifs, run MotifFinder to find motif presence probability}
\usage{
find_motifs_parallel(motif, database = "hocomoco", custom = F,
  seqs = tss_seqs_s)
}
\arguments{
\item{motif}{integer; motif id in object 'hocomoco'}

\item{database}{string; either "hocomoco" (default), or "jaspar"}

\item{custom}{logical; if TRUE (default: FALSE) the motif parameter should be a PWM}

\item{seqs}{charachter vector; DNA sequences (ALL UPPERCASE) to search motif presence in}
}
\value{

}
\description{
For known motifs, run MotifFinder to find motif presence probability
}
\details{
requires the object hocomoco or MotifDb to be loaded (depending on the database set)
}
