% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{gene_expression_pseudotime}
\alias{gene_expression_pseudotime}
\title{Create long data table of gene expression for a subset of genes with pseudotimes}
\usage{
gene_expression_pseudotime(genes)
}
\arguments{
\item{genes}{character vector; Gene Symbols of the genes to compute imputed expression for}
}
\value{
A data.table with variables cell (the cell barcode), PseudoTime, Tsne coordinates,
value (containing imputed expression values), and Gene (a string indicator for which gene each value is for)
}
\description{
Create long data table of gene expression for a subset of genes with pseudotimes
}
\details{
This functions requires the datat data.table and the data matrix to be loaded
}
