% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{gene_list}
\alias{gene_list}
\title{Print list of genes in component and their annotations}
\usage{
gene_list(i, n = 100, factorisation = SDAresults,
  annotations = gene_annotations, order = "absolute")
}
\arguments{
\item{i}{integer or string; number or name of component in 'SDAresults' object}

\item{n}{integer; number of genes to include in the list, default=100}

\item{factorisation}{SDA factorisation object, output of SDAtools::load_results()}

\item{annotations}{data.table with columns gene_symbol, fold_enrichment_fantom, fold_enrichment_gtex,
Human_Orthologue, Inferitility_Gene, chromosome_name and start_position, created in 24_Gene-annotations.Rmd}

\item{order;}{order of ouput list by loading, "absolute", "decreasing" or "increasing"}
}
\value{
print of data.table
}
\description{
Print list of genes in component and their annotations
}
\details{
See also get_top_genes() for a simpler function
}
