% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{get_component_motifs}
\alias{get_component_motifs}
\title{For an SDA component, find motifs /de novo/}
\usage{
get_component_motifs(component, positive, topn = 250, random = FALSE)
}
\arguments{
\item{component}{numeric; component number in results$loadings[[1]]}

\item{positive}{logical; should we look for motifs in the genes with positive loadings or negative}

\item{random}{logical; if TRUE (default: FALSE) 250 random genes are used}
}
\value{
Will save pdf of motifs found, and return list of motifs found
}
\description{
For an SDA component, find motifs /de novo/
}
\details{
Requires tss_seqs_s and results (from SDA load_results) to be loaded
For an SDA
}
