% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{get_top_genes}
\alias{get_top_genes}
\title{Find the top x genes for a component}
\usage{
get_top_genes(component, factorisation = results, n = 20, values = FALSE)
}
\arguments{
\item{component}{integer or string; number or name of component in 'results' object}

\item{factorisation}{SDA factorisation object, output of SDAtools::load_results()}

\item{n}{integer; number of gene symbols to retrieve}

\item{values}{logical; if TRUE (default FALSE) named gene loadings are returned}
}
\value{
charachter vector of gene symbols, or named gene loadings are returned if values=TRUE
}
\description{
Find the top x genes for a component
}
