% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R
\name{GO_enrichment}
\alias{GO_enrichment}
\title{Perform Gene Ontology (Biological Process) enrichment analysis of components}
\usage{
GO_enrichment(component, geneNumber = 250, side = "N",
  factorisation = results, database = Musculus)
}
\arguments{
\item{component}{string; name of component e.g. "V5N" (N meaning negative size loadings)}

\item{geneNumber}{integer; number of top genes to include}

\item{side}{char; should negative of positive values be considered the top}

\item{database}{OrgDb; object of class OrgDb from the package AnnotationDbi}
}
\value{
data.frame with each row a GO term, columns Description, pvalue, Enrichment, etc
}
\description{
Perform Gene Ontology (Biological Process) enrichment analysis of components
}
