% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{go_volcano_plot}
\alias{go_volcano_plot}
\title{Volcano plot of gene ontology enrichment resultd}
\usage{
go_volcano_plot(x = GO_data, component = "V5N", top_n = 30,
  label_size = 3, OR_threshold = 1)
}
\arguments{
\item{x}{R object; GO results}

\item{component}{string; name of component e.g. "V5N" (N meaning negative size loadings)}

\item{top_n}{integer; how many GO terms to label}

\item{label_size}{numeric; size of GO term labels}

\item{OR_threshold}{numeric; threshold on odds ratio of GO terms to label (applied after top_n)}
}
\value{
ggplot2 object
}
\description{
Volcano plot of gene ontology enrichment resultd
}
