% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{imputed_vs_raw}
\alias{imputed_vs_raw}
\title{Plot gene expression through pseudotime (Raw & Imputed)}
\usage{
imputed_vs_raw(genes_tmp, facet = T)
}
\arguments{
\item{facet}{logical; if TRUE (default) each column is a faceted plot,
if FALSE just two plots are returned imputed & raw}

\item{genes}{character vector; Gene Symbols of the genes to compute imputed expression for}
}
\value{
A ggplot2 object
}
\description{
Plot gene expression through pseudotime (Raw & Imputed)
}
\details{
This functions requires the datat data.table and the data matrix to be loaded
}
