% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{load2}
\alias{load2}
\title{Save R Objects, as multiple files in a folder}
\usage{
load2(folder, pattern = ".rds$")
}
\arguments{
\item{folder}{string; path of folder containing objects to be loaded}

\item{pattern}{an optional regular expression. 
Only files which match the regular expression will be loaded.}
}
\description{
Like load but instead of one rds file, loads all rds objects in a folder.
This gives much more flexibility, to add, modify, rename, remove and
share individual objects.
See also save2
}
