% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ordering.R
\name{load_component_orderings}
\alias{load_component_orderings}
\title{Load component orderings in global environ}
\usage{
load_component_orderings()
}
\value{
!! This function modifies (populates) the global environment with objects relating to component names and orders !!
}
\description{
Load component orderings in global environ
}
