% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{load_curve_data}
\alias{load_curve_data}
\title{Load Curve Data}
\usage{
load_curve_data(princurves = principal_curves,
  principal_curve = "df_9")
}
\arguments{
\item{princurves}{list of outputs of princurve()}

\item{principal_curve}{string; name of list entry in principal_curves object, to use for pseudotime line}
}
\value{
data.table with columns for position of curve, annotated with Stage
}
\description{
Load Curve Data
}
