% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{mean_prob_bycomp}
\alias{mean_prob_bycomp}
\title{DEPRECATED, only use if you want custom number of n, For a given component, display binned associations}
\usage{
mean_prob_bycomp(regprobs = regprobs_matrix_custom, plot = TRUE,
  n = 100, factorisation = SDAresults, facet_scales = "free_y")
}
\arguments{
\item{regprobs}{matrix; rows = genes, columns = motif/transcription factors}

\item{plot}{logical; should results be plotted (default), or return matrix?}

\item{n}{int; top n loadings will be used in the calculation of the mean}

\item{factorisation}{list; output from SDAtools::load_results()}
}
\value{
ggplot2 object / matrix (depending on plot parameter)
}
\description{
DEPRECATED, only use if you want custom number of n, For a given component, display binned associations
}
\details{

}
