% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{melt_genes}
\alias{melt_genes}
\title{Create long table combining gene expression and cell metadata}
\usage{
melt_genes(genes, predict = FALSE)
}
\arguments{
\item{genes}{character vector; vector of gene symbols, of which expression values will be plotted over pseudotime}

\item{predict}{logical; if TRUE (default FALSE) imputed/predicted gene expression is used}
}
\value{
data.table with columns 'cell', 'PsuedoTime', 'Tsne1_QC1', 'Tsne2_QC1', 'Gene', and 'Expression'
}
\description{
Create long table combining gene expression and cell metadata
}
\details{
requires 'datat' object to be loaded
}
