% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{Mode}
\alias{Mode}
\title{Find mode of values}
\usage{
Mode(x)
}
\arguments{
\item{x}{numeric vector; values to find Mode of}
}
\value{
numeric value
}
\description{
Find mode of values
}
\details{
Finds most common value in vector
}
