% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{motif_consensus}
\alias{motif_consensus}
\title{Extract matrix of motif occourences from sequence_matrix}
\usage{
motif_consensus(sequence_matrix, motif)
}
\arguments{
\item{sequence_matrix}{output of sequence_matrix function}

\item{motif}{output of MotifFinder}
}
\value{
Charachter matrix (one motif occourence per row, one column per nucleotide) with rownames
}
\description{
Extract matrix of motif occourences from sequence_matrix
}
\details{
From a sequence matrix, extract occourences of the motif into a new sequence matrix
}
