% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{plot_component_regprobs}
\alias{plot_component_regprobs}
\title{plot regprobs within binned components}
\usage{
plot_component_regprobs(component, binned_meanprob,
  facet_scales = "free_y")
}
\arguments{
\item{component}{numeric or string; rowname of `SDAresults$loadings[[1]]` corresponding to a component name or number}
}
\value{
ggplot2 object
}
\description{
plot regprobs within binned components
}
\details{
Requires regprobs_matrix to be loaded.
Calculate mean regulation probability of motif for binned gene loadings from SDA. Plot is faceted by motif.
}
