% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{plot_meanprob_bycomp}
\alias{plot_meanprob_bycomp}
\title{Plot meanprob over components}
\usage{
plot_meanprob_bycomp(binned_meanprob, plot = TRUE, facet_scales = "free_y",
  highlight = NULL)
}
\arguments{
\item{plot}{logical; should results be plotted (default), or return matrix?}

\item{regprobs}{matrix; rows = genes, columns = motif/transcription factors}

\item{n}{int; top n loadings will be used in the calculation of the mean}

\item{factorisation}{list; output from SDAtools::load_results()}
}
\value{
ggplot2 object / matrix (depending on plot parameter)
}
\description{
Plot meanprob over components
}
\details{

}
