% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{plot_motif_matches}
\alias{plot_motif_matches}
\title{For denovo motif and known motif, plot sequence logos}
\usage{
plot_motif_matches(best_denovo, titles = TRUE, yaxis = FALSE)
}
\arguments{
\item{best_denovo}{integer; motif id in object 'hocomoco'}

\item{titles}{logical; if TRUE (default) the logos titles are shown using $Target.Name and $X.Query.ID from best_denovo}

\item{yaxis}{logical; if TRUE (default: FALSE) yaxis is shown}
}
\value{
list of ggplot2 objects
}
\description{
For denovo motif and known motif, plot sequence logos
}
\details{
requires the object results_all
best_denovo is a data.table with column X.Query.ID with value e.g. C13N.I47 matching names of results_all list
also needs orientation etc.

This function is a wrapper for the plot_tomtom_match function from MotifFinder
}
