% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{plot_pseudotime_expression_panel}
\alias{plot_pseudotime_expression_panel}
\title{Plot pseudotime expression panel}
\usage{
plot_pseudotime_expression_panel(data, ncol = 7, title = "Histone Genes",
  gam_k = 5, point_size = 0.2, highlight_reigon = NULL)
}
\arguments{
\item{data}{character vector; vector of gene symbols, of which expression values will be plotted over pseudotime}

\item{ncol}{numeric; number of columns to use for faceting}

\item{title}{string; Title to be used in plot}

\item{gam_k}{numeric; passed to gam formula, controls smoothing}

\item{point_size}{numeric; size of each cell (point) in plot}

\item{highlight_reigon}{numeric vector; vector of size two with min and max pseudotime values between which to highlight in the plot}
}
\value{
ggplot2 object
}
\description{
Plot pseudotime expression panel
}
\details{
requires 'datat' object to be loaded
}
