% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{print_binned_meanprob}
\alias{print_binned_meanprob}
\title{For a given component, display binned associations}
\usage{
print_binned_meanprob(component)
}
\arguments{
\item{component}{numeric or string; rowname of `results$loadings[[1]]` corresponding to a component name or number}
}
\value{
ggplot2 object
}
\description{
For a given component, display binned associations
}
\details{
Requires regprobs_matrix to be loaded.
Calculate mean regulation probability of motif for binned gene loadings from SDA. Plot is faceted by motif.
}
