% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{print_loadings_scores}
\alias{print_loadings_scores}
\title{Plot gene loadings with cell scores underneath}
\usage{
print_loadings_scores(i, max.items = 30)
}
\arguments{
\item{i}{integer or string; number or name of component in 'results' object}

\item{max.items}{integer; number of genes to label in loadings plot}
}
\value{
ggplot2 object
}
\description{
Plot gene loadings with cell scores underneath
}
\details{
requires 'results' and 'rna_locations' objects to be loaded
}
