% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{print_pca}
\alias{print_pca}
\title{Print PCA}
\usage{
print_pca(pc)
}
\arguments{
\item{pc}{string; name of principal compopnent (column of datat), of which cell score values will be used to colour points (cells)}
}
\value{
ggplot2 object
}
\description{
Print PCA
}
\details{
requires PCA object to be bound to datat object and 'datat' object to be loaded
}
