% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{print_tsne}
\alias{print_tsne}
\title{Plot tSNE}
\usage{
print_tsne(i, flip = FALSE, predict = FALSE, curve = FALSE,
  point_size = 0.25, log = FALSE, principal_curve = "df_9",
  curve_width = 0.5)
}
\arguments{
\item{i}{; either a numeric value corresponding to the component to plot, a gene name,
or the name of a variable stored in datat such as "group", "PseudoTime", or "library_size"}

\item{flip}{logical; should the cell scores be multiplied by -1 (score/gene loading sign combination are arbitrary like PCA scores).}

\item{predict}{logical; Should gene expression be the predicted (imputed) values or raw normalised}

\item{curve}{logical; Should principal curve (pseudotime) by plotted on top of the tSNE plot}

\item{point_size}{numeric; size of point}

\item{principal_curve}{string; name of list entry in principal_curves object, to use for pseudotime line}
}
\value{
The Y matrix from Rtsne output, rotated by angle
}
\description{
Plot tSNE
}
