% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ordering.R
\name{ptorder}
\alias{ptorder}
\title{Order Components by PseudoTime}
\usage{
ptorder(component, cell_metadata = datat, threshold = 2)
}
\arguments{
\item{component}{string; Component name as in datat}
}
\value{
weighted mean of pseudotime for a given component
}
\description{
Order Components by PseudoTime
}
\details{
The order is determined by using a weighted mean of the pseudotime values, where the weights are the cell scores of 
the component. Only cells with an absolute cell score of greater than 2 contribute to the mean.
}
