% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{reverse_normalisation}
\alias{reverse_normalisation}
\title{Reverse DGE Normalisation}
\usage{
reverse_normalisation(dge, sign = TRUE, cells = cell_subset,
  lib_correction = lib_size_correction_factor, standard_dev = sds)
}
\arguments{
\item{dge}{numeric matrix; rows are cells, columns are genes}

\item{sign}{logical; if TRUE (default) the sign of the values is retained (negatives are still negative)}

\item{cells}{character vector; strings corresponding to cell names to include}

\item{lib_correction}{numeric vector; size correction factor used for each cell}

\item{standard_dev}{numeric vector; standard deviations used to scale each gene}
}
\value{
Matrix of cells by genes, with values on the count scale
}
\description{
Reverse DGE Normalisation
}
\details{
Reverses the transformation of dropsim::normaliseDGE() to get back to a counts scale
}
