% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{rotate_matrix}
\alias{rotate_matrix}
\title{Generate rotation matrix / rotate a matrix}
\usage{
rotate_matrix(object, angle, rotation = FALSE)
}
\arguments{
\item{object}{matrix; matrix with two columns}

\item{angle}{numeric; degrees to rotate the tSNE/UMAP/PCA embedding by}

\item{rotation}{logical; if TRUE returns rotation matrix instead of the rotated matrix (default: FALSE)}
}
\value{
Rotated matrix
}
\description{
Generate rotation matrix / rotate a matrix
}
