% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{rotate_SDA}
\alias{rotate_SDA}
\title{Rotate SDA factorisation}
\usage{
rotate_SDA(X = results9_WT, reference = results1)
}
\arguments{
\item{X}{SDA results list to be rotated}

\item{reference}{SDA results list to for X to be rotated towards}
}
\value{
SDA results list, but with the gene loadings and scores matrices rotated by procrustes
}
\description{
Rotate SDA factorisation
}
