% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{rotate_tsne}
\alias{rotate_tsne}
\title{Rotate a tSNE embedding about the origin}
\usage{
rotate_tsne(tsne, angle)
}
\arguments{
\item{tsne}{object; output of Rtsne::Rtsne()}

\item{angle}{numeric; degrees to rotate the tSNE embedding by}
}
\value{
The Y matrix from Rtsne output, rotated by angle
}
\description{
Rotate a tSNE embedding about the origin
}
