% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{save2}
\alias{save2}
\title{Save R Objects, as multiple files in a folder}
\usage{
save2(folder = "", list = NULL, compress = FALSE)
}
\arguments{
\item{folder}{string; path of folder in which to save objects}

\item{list}{character vector; names of objects to be saved.
Defaults to all objects in global environment (not functions).}

\item{compress}{logical; specifying whether saving to a named file is to use "gzip" compression,
or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used.}
}
\description{
Like save but instead of one rds file, creates one file per object
This gives much more flexibility, to add, modify, rename, remove and
share individual objects.
}
\details{
See also load2
}
