% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{sequence_matrix}
\alias{sequence_matrix}
\title{Create matrix representation of sequences with highlighted motif}
\usage{
sequence_matrix(motif)
}
\arguments{
\item{motif}{output of MotifFinder}
}
\value{
Charachter matrix (one sequence per row, one column per nucleotide) with rownames
}
\description{
Create matrix representation of sequences with highlighted motif
}
\details{
From the output of MotifFinder, creates matrix of sequences, with the motif found highlighted using lowercase nucleotide letters
}
