% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{sort_matrix}
\alias{sort_matrix}
\title{Sort matrix to highest correlations are along the diagonal}
\usage{
sort_matrix(mat = cross_cor, order1 = max_cor_per_compoent)
}
\arguments{
\item{mat}{numeric matrix to be sorted}

\item{order1}{order of rows}
}
\value{
numeric matrix
}
\description{
Sort matrix to highest correlations are along the diagonal
}
\details{
For a given row order, sorts the columns so that each row
has the highest correlation (in a greedy fashion)
}
