% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{split_loadings}
\alias{split_loadings}
\title{Split SDA components into positive and negative components}
\usage{
split_loadings(loadings = results$loadings[[1]])
}
\arguments{
\item{loadings}{numeric matrix; components vs gene matrix of loadings, e.g. `SDAtools::load_results()$loadings[[1]]`}
}
\value{
loadings numeric matrix, with twice as many columns
}
\description{
Split SDA components into positive and negative components
}
\details{

}
