% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{ternaryColours}
\alias{ternaryColours}
\title{Plot tricolor tSNE}
\usage{
ternaryColours(df, variables, shift = FALSE)
}
\arguments{
\item{df;}{string vector of two or three gene names}

\item{variables;}{string vector of two or three variable
(gene) names present in df to create colours from}

\item{shift;}{logical, should the values be shifted so most 
negative becomes 0, for truely bi-signed variables (default: F)}
}
\value{
vector of colour rbg hex codes
}
\description{
Plot tricolor tSNE
}
