% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tss.R
\name{export_all_pwms}
\alias{export_all_pwms}
\title{Export PWMs in MEME format}
\usage{
export_all_pwms(results_all, file)
}
\arguments{
\item{results_all}{list of outputs from MotifFinder}

\item{file}{string; name of file to save PWMs to}
}
\value{
nothing, results written to file
}
\description{
Export PWMs in MEME format
}
\details{
Export list of MotifFinder results to MEME format file, e.g. for input into TOMTOM
}
