% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{gene_expression_pseudotime}
\alias{gene_expression_pseudotime}
\title{Create long data table of gene expression for a subset of genes with pseudotimes}
\usage{
gene_expression_pseudotime(genes, factorisation = SDAresults,
  cell_metadata = cell_data)
}
\arguments{
\item{genes}{character vector; Gene Symbols of the genes to compute imputed expression for}

\item{factorisation}{SDA factorisation object, output of SDAtools::load_results()}

\item{cell_metadata}{data.table with columns cell, Tsne1_QC1, Tsne2_QC2, and components V1, V2 etc.}
}
\value{
A data.table with variables cell (the cell barcode), PseudoTime, Tsne coordinates,
value (containing imputed expression values), and Gene (a string indicator for which gene each value is for)
}
\description{
Create long data table of gene expression for a subset of genes with pseudotimes
}
