% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{manhatten_plot}
\alias{manhatten_plot}
\title{Plot Manhatten Style Enrichment for each component}
\usage{
manhatten_plot(enrichments, topn = 1, repel_force = 1,
  legend_position = c(0.85, 0.8))
}
\arguments{
\item{enrichments}{result of the component_enrichment() function}

\item{topn}{int; how many components should be labeled}

\item{repel_force;}{degree of force to move points apart form each other}

\item{legend_position;}{numeric vector of length 2, giving positions
between 0 and 1 where the legend should be placed  within the plot}
}
\value{
A ggplot object
}
\description{
Plot Manhatten Style Enrichment for each component
}
\details{

}
