% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{plotCellAUC}
\alias{plotCellAUC}
\title{Plot ROC curve for imputation rankings}
\usage{
plotCellAUC(i, mt)
}
\arguments{
\item{i}{integer or string; number or name of cell in the matrices}

\item{mt}{named list of  gene rank by cell matricies, cumulative sum of predicted for each cell}
}
\value{
ggplot2 object
}
\description{
Plot ROC curve for imputation rankings
}
