% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{plotCellAUC}
\alias{plotCellAUC}
\title{Plot ROC curve for imputation rankings}
\usage{
plotCellAUC(i)
}
\arguments{
\item{i}{integer or string; number or name of cell}
}
\value{
ggplot2 object
}
\description{
Plot ROC curve for imputation rankings
}
\details{
! nb this function requires the objects cumsum_predict, cumsum_train, and cumsum_mean to be loaded.
}
