% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{print_gene_list}
\alias{print_gene_list}
\title{Print list of genes in component and their annotations}
\usage{
print_gene_list(i)
}
\arguments{
\item{i}{integer or string; number or name of component in 'results' object}
}
\value{
print of data.table
}
\description{
Print list of genes in component and their annotations
}
\details{
requires 'results', 'fantom_summary_subset', 'gtex_summary' and 'infertility_genes' objects to be loaded
}
