% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{print_raw_tsne}
\alias{print_raw_tsne}
\title{Print tsne based on raw expression rather than SDA scores}
\usage{
print_raw_tsne(gene, expression_matrix = data,
  cell_metadata = cell_data)
}
\arguments{
\item{gene}{string; gene symbol, of which expression values will be used to colour points (cells)}

\item{expression_matrix}{cell by gene expressionn matrix}

\item{cell_metadata}{data.table with columns Tsne1, Tsne2, and PCs}
}
\value{
ggplot2 object
}
\description{
Print tsne based on raw expression rather than SDA scores
}
