% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{sda_predict}
\alias{sda_predict}
\title{Compute Imputed Expression values}
\usage{
sda_predict(genes, factorisation = results, name_extension = "")
}
\arguments{
\item{genes}{character vector; Gene Symbols of the genes to compute imputed expression for}

\item{factorisation}{SDA factorisation object, output of SDAtools::load_results()}

\item{name_extension}{string; will be appended to any gene names given (so resulting table could be joined with original values)}
}
\value{
A data.table with variables cell (the cell barcode) and a column for each gene containing the imputed expression values
}
\description{
Compute Imputed Expression values
}
