% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_single_cell.R
\name{single_component_enrichment}
\alias{single_component_enrichment}
\title{Calculate enrichment for a set of genes in a single component}
\usage{
single_component_enrichment(component, genes = Mybl1_genes,
  factorisation = results, pos = T, return_genes = F, threshold = 200,
  bg_genes = NULL)
}
\arguments{
\item{component}{number or name of component}

\item{genes}{character vector of gene set to look for enrichment for}

\item{factorisation}{SDA factorisation object, output of SDAtools::load_results()}

\item{pos}{logical; if TRUE (default) the positive gene loadings are ranked highest}

\item{return_genes}{logical; if TRUE (default: FALSE) ranking of genes in this component is returned,
if false, output of fisher.test is returned}

\item{threshold}{numeric; how many genes should be included in the top genes list used to calculate enrichment}

\item{bg_genes}{which genes to use as the 'background' default uses all genes in the results object from SDA}
}
\value{

}
\description{
Calculate enrichment for a set of genes in a single component
}
\details{
Calculate p value of enrichemnt (fishers test) in a component, given a set of genes
requires the results object from SDA to be loaded
}
